+++ src/web/common/util/time.py
@@ -1,0 +1,7 @@
+from datetime import datetime
+from django.conf import settings
+from zoneinfo import ZoneInfo
+
+
+def now():
+    return datetime.now(ZoneInfo(settings.TIME_ZONE))