+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Make the "Finish now?" button on session list work
 * Hide navigation bar items if there are no games/purchases/sessions
 * Set default version to "git-main" to indicate development environment
 * Add homepage, link to it from the logo